refactor(types): split up article content types
