+++ src/types/article/contentBase.ts
@@ -1,0 +1,4 @@
+export type ArticleContentBase = {
+  id: string;
+  data: string;
+};
