+++ src/types/article/contentBody.ts
@@ -1,0 +1,9 @@
+import { type BodyProps } from "@/components/ui/Body/interfaces";
+
+import { type ArticleContentBase } from "./contentBase";
+
+export interface ArticleContentBody
+  extends ArticleContentBase,
+    Omit<BodyProps, "children"> {
+  type: "body";
+}
