+++ src/types/article/contentList.ts
@@ -1,0 +1,9 @@
+import { type ListProps } from "@/components/ui/List/interfaces";
+
+import { type ArticleContentBase } from "./contentBase";
+
+export interface ArticleContentList
+  extends Omit<ArticleContentBase, "data">,
+    ListProps {
+  type: "list";
+}
