+++ src/types/article/index.ts
@@ -1,0 +1,9 @@
+import { type ArticleProps } from "@/components/ui/Article/interfaces";
+
+import { type ArticleContent } from "./content";
+
+export type ArticleData = {
+  id: string;
+  gap?: ArticleProps["gapVariant"];
+  content: ArticleContent[];
+};
